build: Updating dependencies
composer:
* mediawiki/mediawiki-codesniffer: 49.0.0 → 50.0.0
* mediawiki/mediawiki-phan-config: 0.18.0 → 0.20.0

npm:
* svgo: 3.2.0 → 3.3.3
* basic-ftp: 5.0.5 → 5.2.0
  * https://github.com/advisories/GHSA-5rq4-664w-9x2c
* fast-xml-parser: 5.3.7 → 5.5.5
  * https://github.com/advisories/GHSA-fj3w-jwp8-x2g3
* flatted: 3.3.1 → 3.4.1
  * https://github.com/advisories/GHSA-25h7-pfq9-p65f
* undici: 6.23.0, 7.18.2 → 6.24.0, 7.24.1
  * https://github.com/advisories/GHSA-2mjp-6q6p-2qxm
  * https://github.com/advisories/GHSA-4992-7rv2-5pvq
  * https://github.com/advisories/GHSA-f269-vfmq-vjvj
  * https://github.com/advisories/GHSA-phc3-fgpg-7m6h
  * https://github.com/advisories/GHSA-v9p9-hfj2-hcw8
  * https://github.com/advisories/GHSA-vrm6-8vpv-qv8q

+++ includes/specials/CentralNotice.php
@@ -848,7 +848,7 @@ class CentralNotice extends UnlistedSpecialPage {
 				// Handle user bucketing setting for campaign
 				$numCampaignBuckets = min( $request->getInt( 'buckets', 1 ),
 					$noticeNumberOfBuckets );
-				$numCampaignBuckets = pow( 2, floor( log( $numCampaignBuckets, 2 ) ) );
+				$numCampaignBuckets = (int)pow( 2, floor( log( $numCampaignBuckets, 2 ) ) );
 
 				Campaign::setNumericCampaignSetting(
 					$notice,
